git commit -m "auth files"

+++ app/src/main/java/in/mercuryai/emptyproject/presentation/util/SnackbarEvent.kt
@@ -4,5 +4,5 @@ import androidx.compose.material3.SnackbarDuration
 
 data class SnackbarEvent(
     val message: String,
-    val duration: SnackbarDuration = SnackbarDuration.Short
+    val duration: SnackbarDuration = SnackbarDuration.Long
 )
